Add OSX Support to Makefile.
On OSX, the functionality of ncursesw is bundled into ncurses. However,
the system ncurses library is out of date and does not have all the
functions used here. Using homebrew, one can install a later version
of the ncurses library that has the functions needed.

In this Makefile, the homebrew paths for ncurses are explicitely defined in the
CFLAGS and LIBS variables.